Crate simple_stats

A simple statistics library

Heavily inspired by the python standard library statistics module.

Functions

mean
The mean is the sum of a collection of numbers divided by the number of numbers in the collection. (reference)[http://en.wikipedia.org/wiki/Arithmetic_mean]

median
The median is the number separating the higher half of a data sample, a population, or a probability distribution, from the lower half (reference)[http://en.wikipedia.org/wiki/Median)

population_standard_deviation

population_variance
(Population variance)[http://en.wikipedia.org/wiki/Variance#Population_variance]

standard_deviation
Standard deviation is a measure that is used to quantify the amount of variation or dispersion of a set of data values. (reference)[http://en.wikipedia.org/wiki/Standard_deviation]

variance
(Sample variance)[http://en.wikipedia.org/wiki/Variance#Sample_variance]