Function simple_stats::median

pub fn median<T>(v: &[T]) -> T where T: Copy + Num + NumCast + PartialOrd

The median is the number separating the higher half of a data sample, a population, or a probability distribution, from the lower half (reference)[http://en.wikipedia.org/wiki/Median)